task 11: declared r & l as size_t

+++ 11-binary_tree_size.c
@@ -8,7 +8,7 @@
  */
 size_t binary_tree_size(const binary_tree_t *tree)
 {
-	int r, l;
+	size_t r, l;
 
 	if (tree == NULL)
 		return (0);
